move components out of container dir and create Icon components
