+++ src/components/styled/PanelIcons.js
@@ -1,0 +1,9 @@
+import React from "react"
+import styled from "styled-components"
+
+const PanelIcons = styled.div`
+  display: flex;
+  align-items: flex-end;
+`
+
+export default PanelIcons
