+++ src/components/styled/StyledIcon.js
@@ -1,0 +1,7 @@
+import React from "react"
+import styled from "styled-components"
+
+const StyledIcon = styled.button`
+`
+
+export default StyledIcon
